+++ ChangeLog.md
@@ -1,6 +1,6 @@
 # ChangeLog for Related
 
 ## Unreleased
-
+- NEW: Script reloadCustomMultientities - add param "modules" to limit modules to reload - 1.2.0 - *23/12/2024*
 - NEW: Script reloadCustomMultientities to reload only custom module for each entity - 1.1.0 - *11/03/2024*
 - no ChangeLog up to this point